Implementation of application profile abstraction.
The goal of this class is to abstract FrequencyData produced by grinder. We plan to provide query for the kind of question needed by PGO.

IsBlockWarm, etc,...

And, in addition, we now have global metrics like global_temperature and the percentile of a block (relative to the global temperature).

More metrics are to come (i.e. devirtualisation), and they will be abstract by this class.

R=chrisha@chromium.org, sebmarchand@chromium.org

BUG=

Review URL: https://codereview.appspot.com/20400043

git-svn-id: http://sawbuck.googlecode.com/svn/trunk@1878 15e8cca8-e42c-11de-a347-f34a4f72eb7d