Add instrumentation subsampling support to ASan instrumenter.
This allows a randomly subsampled fraction of read/writes to be instrumented rather than simply instrumenting all of them. Setting this to zero means that no instructions are instrumented, but that functions will still be intercepted. Combined with the newly enabled quarantine checksumming this acts as an extremely low cost use-after-free detector.

BUG=
R=etienneb@chromium.org, sebmarchand@chromium.org

Review URL: https://codereview.appspot.com/71650043

git-svn-id: http://sawbuck.googlecode.com/svn/trunk@2059 15e8cca8-e42c-11de-a347-f34a4f72eb7d