Expose the PDB hash function.
The hash function used in the named streams table of a PDB file is also
used to generate the "public symbols" stream.

This CL moves the hash function out of the anonymous namespace of
pdb_util.cc so that it can be used to generate a "public symbols"
stream.

BUG=
R=chrisha@chromium.org, etienneb@chromium.org, sebmarchand@chromium.org

Review URL: https://codereview.appspot.com/115370043

git-svn-id: http://sawbuck.googlecode.com/svn/trunk@2204 15e8cca8-e42c-11de-a347-f34a4f72eb7d